Problem 58 Spiral primes BLUE

+++ net.projecteuler.solutions/src/test/java/problem/no21to30/Problem28Test.java
@@ -3,8 +3,6 @@ package problem.no21to30;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import java.math.BigInteger;
-
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class Problem28Test {
@@ -16,8 +14,6 @@ class Problem28Test {
             "961, 11"
     })
     void sumOfDiagonalsInAGridOfSize1x1(int expected, int gridSize) {
-        assertEquals(BigInteger.valueOf(expected), Problem28.sumOfDiagonalsInAGridOfSizeNxN(gridSize));
+        assertEquals(expected, Problem28.sumOfDiagonalsInAGridOfSizeNxN(gridSize));
     }
-
-
 }
